requirements.txt
supabase==2.4.0
requests==2.31.0
PyJWT==2.8.0
cryptography==42.0.0
python-dotenv==1.0.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
google-ads==24.1.0
google-cloud-bigquery==3.17.2
